Move web, state, auth into cycls/app/ subpackage
Extract sessions + files endpoints into state.py, move web server
and auth into cycls/app/ directory. Extract session_id from message
history for multi-turn continuity.
